move pebble/cache to pebble/internal/cache
Users will have to adjust their code, but there shouldn't be many such
users yet, and that fixup seems preferrable to leaving in a backwards
compatible shim.
