+++ internal/cache/alloc_test.go
